add bigint cast to replay_lag_bytes in bdr check queries

+++ query_files/bdr-replication-slot-lag-time.sql
@@ -6,7 +6,7 @@ threshold:
 --- */
 select slot_name
      , extract('epoch' from replay_lag)::int as replay_lag_seconds
-     , replay_lag_bytes::int
+     , replay_lag_bytes::bigint
      , replay_lag_size
   from bdr.node_slots
  where slot_name != 'bdr_wmx_rail9f667b5_wmx_api'
